tiny cleanups
