+++ script/build.js
@@ -1,11 +1,11 @@
 ({
-  baseUrl: 'src',
-  mainConfigFile: 'src/script.js',
+  baseUrl: '../src',
+  mainConfigFile: '../src/script.js',
   preserveLicenseComments: true,
   wrap: false,
   name: '../node_modules/almond/almond',
   include:'script',
   insertRequire:['script'],
-  out: 'dist/src/script.js',
+  out: '../dist/src/script.js',
   optimize: 'uglify2'
 })
